fix: 배팅 금액 출력문에 소숫점이 나오는 문제 수정

+++ src/main/java/blackjack/BlackjackGame.java
@@ -93,7 +93,7 @@ public class BlackjackGame {
         );
     }
 
-    private List<PlayerGameResult> toGameResults(Map<Player, Double> countGameResult) {
+    private List<PlayerGameResult> toGameResults(Map<Player, Integer> countGameResult) {
         return countGameResult.keySet().stream()
             .map(player -> new PlayerGameResult(player.getName(), countGameResult.get(player)))
             .collect(Collectors.toList());
